move action const into separate file and add getTagKey() helper
